+++ CHANGELOG.md
@@ -8,6 +8,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Added
 
+- new models:
+  - [MeliusNet](bitorch/models/meliusnet.py)
+  - [BinaryDenseNet](bitorch/models/densenet.py)
 - simple example script for MNIST
 - support for integration of bitorch's inference engine for the following layers
   - QLinear
@@ -29,6 +32,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - code is now formatted with the black code formatter
 - using PyTorch's implementation of RAdam
 - renamed the `bitwidth` attribute of quantization functions to `bit_width`
+- moved the image datasets out of the bitorch core package into the image classification example
 
 ### Fixed